% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf-get.R
\name{cf_GET}
\alias{cf_GET}
\title{Cloudflare bypass version of httr::GET}
\usage{
cf_GET(cf_url, cf_agent = NULL, ...)
}
\arguments{
\item{cf_url}{the URL you want}

\item{cf_agent}{user agent you want to use. if \code{NULL} a valid one is chosen at random}

\item{...}{other params passed to all \code{httr::GET} calls (other headers, verbose, etc)}
}
\value{
\code{httr::response object}
}
\description{
If this ceases to work it likely means that Cloudflare changed the challenge code.
Please \href{https://github.com/hrbrmstr/cfhttr/issues}{file an issue} so I can redo the
workaround.
}
\note{
This function pick a user agent at random if one isn't passed in
}
\examples{
res <- cf_GET("https://www.bitmarket.pl/graphs/BTCPLN/90m.json")
str(httr::content(res, as="parsed"))
}
